+++ include/vkl/scene_tree/vkl_mesh.hpp
@@ -146,7 +146,7 @@ void VklMesh<VertexType, IndexType, BoxType>::allocDescriptorSets(VklDescriptorS
         if (not key.uniformDescriptors.empty()) {
             bufferInfos.reserve(uniformBuffers[key].size() * 5);
 
-            for (int i = 0; i < uniformBuffers[key].size(); i++) {
+            for (size_t i = 0; i < uniformBuffers[key].size(); i++) {
                 uniformBuffers[key][i] = std::move(std::make_unique<VklBuffer>(
                     device_, key.uniformDescriptors.front().size, 1, VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT,
                     VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT));
